Move <Menu> to app level
This is a bit of a conceptual one. If you think about it, the menu system is not a child of the pages (which was how you had architected it). It is a peer. What I did here was move it to the app level. You will notice that nothing 'changed' visually (except some minor styling issues). The key difference is that now, the menu component only renders once (when the app component renders), rather than each time the page changes.