Add npm lint task
This small change makes it so you can do `npm run lint` from the terminal, and see the current linting errors. This task can be connected to pre commit hooks in your git configuration, making it so that you cannot commit code that doesn't lint. That is fairly extreme (and can be annoying when you just want to get something up and running quickly), but having the linting task seperated like this is useful.