Fix linting issues
Here, I fix the other linting issue that you were disabling previously. It looks like you installed the PropTypes library, but then decided against implementing it! It can be difficult to get started with, but hopefully you can see from the way I did it that it is fairly intuitive. Please let me know if you struggle to know how to use it!