Update <Home> animation to match <Page>
Here I updated the home animation to be more consistent with Page. Notice that I moved the variable holding the variants outside of the component definition. This small optimisation means that the variable is not redeclared each time the component renders (you could achieve a similar result by memoizing the value using useMemo). This sort of thing becomes important for performance when your components get more complicated.